+++ app/scripts/controllers/network.js
@@ -1,72 +1,42 @@
+const assert = require('assert')
 const EventEmitter = require('events')
-const MetaMaskProvider = require('web3-provider-engine/zero.js')
 const ObservableStore = require('obs-store')
 const ComposedStore = require('obs-store/lib/composed')
 const extend = require('xtend')
 const EthQuery = require('eth-query')
+const createEthRpcClient = require('eth-rpc-client')
+const createEventEmitterProxy = require('../lib/events-proxy.js')
+const createObjectProxy = require('../lib/obj-proxy.js')
 const RPC_ADDRESS_LIST = require('../config.js').network
 const ALTERNATIVE_CHAINS = require('../config.js').networkIdAlterantiveChains
 const DEFAULT_RPC = RPC_ADDRESS_LIST['rinkeby']
 
 module.exports = class NetworkController extends EventEmitter {
+
   constructor (config) {
     super()
-    this.networkStore = new ObservableStore('loading')
     config.provider.rpcTarget = this.getRpcAddressForType(config.provider.type, config.provider)
+    this.networkStore = new ObservableStore('loading')
     this.providerStore = new ObservableStore(config.provider)
     this.store = new ComposedStore({ provider: this.providerStore, network: this.networkStore })
-    this._providerListeners = {}
+    this.providerProxy = createObjectProxy()
+    this.blockTrackerProxy = createEventEmitterProxy()
 
     this.on('networkDidChange', this.lookupNetwork)
-    this.providerStore.subscribe((state) => this.switchNetwork({rpcUrl: state.rpcTarget}))
   }
 
-  get provider () {
-    return this._proxy
-  }
-
-  set provider (provider) {
-    this._provider = provider
-  }
-
-  initializeProvider (opts, providerContructor = MetaMaskProvider) {
-    this.providerInit = opts
-    this._provider = providerContructor(opts)
-    this._proxy = new Proxy(this._provider, {
-      get: (obj, name) => {
-        if (name === 'on') return this._on.bind(this)
-        return this._provider[name]
-      },
-      set: (obj, name, value) => {
-        this._provider[name] = value
-        return value
-      },
-    })
-    this.provider.on('block', this._logBlock.bind(this))
-    this.provider.on('error', this.verifyNetwork.bind(this))
-    this.ethQuery = new EthQuery(this.provider)
+  initializeProvider (_providerParams) {
+    this._baseProviderParams = _providerParams
+    const rpcUrl = this.getCurrentRpcAddress()
+    this._configureStandardClient({ rpcUrl })
+    this.blockTrackerProxy.on('block', this._logBlock.bind(this))
+    this.blockTrackerProxy.on('error', this.verifyNetwork.bind(this))
+    this.ethQuery = new EthQuery(this.providerProxy)
     this.lookupNetwork()
-    return this.provider
   }
-
-  switchNetwork (providerInit) {
-    this.setNetworkState('loading')
-    const newInit = extend(this.providerInit, providerInit)
-    this.providerInit = newInit
-
-    this._provider.removeAllListeners()
-    this._provider.stop()
-    this.provider = MetaMaskProvider(newInit)
-    // apply the listners created by other controllers
-    Object.keys(this._providerListeners).forEach((key) => {
-      this._providerListeners[key].forEach((handler) => this._provider.addListener(key, handler))
-    })
-    this.emit('networkDidChange')
-  }
-
 
   verifyNetwork () {
-  // Check network when restoring connectivity:
+    // Check network when restoring connectivity:
     if (this.isNetworkLoading()) this.lookupNetwork()
   }
 
@@ -100,6 +70,7 @@ module.exports = class NetworkController extends EventEmitter {
       type: 'rpc',
       rpcTarget: rpcUrl,
     })
+    this._switchNetwork({ rpcUrl })
   }
 
   getCurrentRpcAddress () {
@@ -108,10 +79,16 @@ module.exports = class NetworkController extends EventEmitter {
     return this.getRpcAddressForType(provider.type)
   }
 
-  setProviderType (type) {
+  async setProviderType (type) {
+    assert(type !== 'rpc', `NetworkController.setProviderType - cannot connect by type "rpc"`)
+    // skip if type already matches
     if (type === this.getProviderConfig().type) return
+    // lookup rpcTarget for typecreateMetamaskProvider
     const rpcTarget = this.getRpcAddressForType(type)
-    this.providerStore.updateState({type, rpcTarget})
+    assert(rpcTarget, `NetworkController - unknown rpc address for type "${type}"`)
+    // update connectioncreateMetamaskProvider
+    this.providerStore.updateState({ type, rpcTarget })
+    this._switchNetwork({ rpcUrl: rpcTarget })
   }
 
   getProviderConfig () {
@@ -127,14 +104,39 @@ module.exports = class NetworkController extends EventEmitter {
     return ALTERNATIVE_CHAINS[type]
   }
 
+  //
+  // Private
+  //
+
+  _switchNetwork (providerParams) {
+    this.setNetworkState('loading')
+    this._configureStandardClient(providerParams)
+    this.emit('networkDidChange')
+  }
+
+  _configureStandardClient(_providerParams) {
+    const providerParams = extend(this._baseProviderParams, _providerParams)
+    const client = createEthRpcClient(providerParams)
+    this._setClient(client)
+  }
+
+  _setClient (newClient) {
+    // teardown old client
+    const oldClient = this._currentClient
+    if (oldClient) {
+      oldClient.blockTracker.stop()
+      // asyncEventEmitter lacks a "removeAllListeners" method
+      // oldClient.blockTracker.removeAllListeners
+      oldClient.blockTracker._events = {}
+    }
+    // set as new provider
+    this._currentClient = newClient
+    this.providerProxy.setTarget(newClient.provider)
+    this.blockTrackerProxy.setTarget(newClient.blockTracker)
+  }
+
   _logBlock (block) {
     log.info(`BLOCK CHANGED: #${block.number.toString('hex')} 0x${block.hash.toString('hex')}`)
     this.verifyNetwork()
-  }
-
-  _on (event, handler) {
-    if (!this._providerListeners[event]) this._providerListeners[event] = []
-    this._providerListeners[event].push(handler)
-    this._provider.on(event, handler)
   }
 }