+++ app/scripts/lib/pending-tx-tracker.js
@@ -1,7 +1,5 @@
 const EventEmitter = require('events')
 const EthQuery = require('ethjs-query')
-const sufficientBalance = require('./util').sufficientBalance
-const RETRY_LIMIT = 3500 // Retry 3500 blocks, or about 1 day.
 /*
 
   Utility class for tracking the transactions as they
@@ -13,7 +11,6 @@ const RETRY_LIMIT = 3500 // Retry 3500 blocks, or about 1 day.
   requires a: {
     provider: //,
     nonceTracker: //see nonce tracker,
-    getBalnce: //(address) a function for getting balances,
     getPendingTransactions: //() a function for getting an array of transactions,
     publishTransaction: //(rawTx) a async function for publishing raw transactions,
   }
@@ -25,11 +22,12 @@ module.exports = class PendingTransactionTracker extends EventEmitter {
     super()
     this.query = new EthQuery(config.provider)
     this.nonceTracker = config.nonceTracker
-
-    this.getBalance = config.getBalance
+    // default is one day
+    this.retryTimePeriod = config.retryTimePeriod || 86400000
     this.getPendingTransactions = config.getPendingTransactions
+    this.getCompletedTransactions = config.getCompletedTransactions
     this.publishTransaction = config.publishTransaction
-    this.giveUpOnTransaction = config.giveUpOnTransaction
+    this._checkPendingTxs()
   }
 
   //  checks if a signed tx is in a block and
@@ -44,18 +42,18 @@ module.exports = class PendingTransactionTracker extends EventEmitter {
       if (!txHash) {
         const noTxHashErr = new Error('We had an error while submitting this transaction, please try again.')
         noTxHashErr.name = 'NoTxHashError'
-        this.emit('txFailed', txId, noTxHashErr)
+        this.emit('tx:failed', txId, noTxHashErr)
         return
       }
 
 
       block.transactions.forEach((tx) => {
-        if (tx.hash === txHash) this.emit('txConfirmed', txId)
+        if (tx.hash === txHash) this.emit('tx:confirmed', txId)
       })
     })
   }
 
-  queryPendingTxs ({oldBlock, newBlock}) {
+  queryPendingTxs ({ oldBlock, newBlock }) {
     // check pending transactions on start
     if (!oldBlock) {
       this._checkPendingTxs()
@@ -76,6 +74,9 @@ module.exports = class PendingTransactionTracker extends EventEmitter {
       Dont marked as failed if the error is a "known" transaction warning
       "there is already a transaction with the same sender-nonce
       but higher/same gas price"
+
+      Also don't mark as failed if it has ever been broadcast successfully.
+      A successful broadcast means it may still be mined.
       */
       const errorMessage = err.message.toLowerCase()
       const isKnownTx = (
@@ -92,63 +93,67 @@ module.exports = class PendingTransactionTracker extends EventEmitter {
       // ignore resubmit warnings, return early
       if (isKnownTx) return
       // encountered real error - transition to error state
-      this.emit('txFailed', txMeta.id, err)
+      txMeta.warning = {
+        error: errorMessage,
+        message: 'There was an error when resubmitting this transaction.',
+      }
+      this.emit('tx:warning', txMeta, err)
     }))
   }
 
   async _resubmitTx (txMeta) {
-    const address = txMeta.txParams.from
-    const balance = this.getBalance(address)
-    if (balance === undefined) return
-    if (!('retryCount' in txMeta)) txMeta.retryCount = 0
-
-    if (txMeta.retryCount > RETRY_LIMIT) {
-      return this.giveUpOnTransaction(txMeta.id)
-    }
-
-    // if the value of the transaction is greater then the balance, fail.
-    if (!sufficientBalance(txMeta.txParams, balance)) {
-      const insufficientFundsError = new Error('Insufficient balance during rebroadcast.')
-      this.emit('txFailed', txMeta.id, insufficientFundsError)
-      log.error(insufficientFundsError)
-      return
+    if (Date.now() > txMeta.time + this.retryTimePeriod) {
+      const hours = (this.retryTimePeriod / 3.6e+6).toFixed(1)
+      const err = new Error(`Gave up submitting after ${hours} hours.`)
+      return this.emit('tx:failed', txMeta.id, err)
     }
 
     // Only auto-submit already-signed txs:
     if (!('rawTx' in txMeta)) return
 
-    // Increment a try counter.
-    txMeta.retryCount++
     const rawTx = txMeta.rawTx
-    return await this.publishTransaction(rawTx)
+    const txHash = await this.publishTransaction(rawTx)
+
+    // Increment successful tries:
+    this.emit('tx:retry', txMeta)
+    return txHash
   }
 
   async _checkPendingTx (txMeta) {
     const txHash = txMeta.hash
     const txId = txMeta.id
+
     // extra check in case there was an uncaught error during the
     // signature and submission process
     if (!txHash) {
       const noTxHashErr = new Error('We had an error while submitting this transaction, please try again.')
       noTxHashErr.name = 'NoTxHashError'
-      this.emit('txFailed', txId, noTxHashErr)
+      this.emit('tx:failed', txId, noTxHashErr)
       return
     }
+
+    // If another tx with the same nonce is mined, set as failed.
+    const taken = await this._checkIfNonceIsTaken(txMeta)
+    if (taken) {
+      const nonceTakenErr = new Error('Another transaction with this nonce has been mined.')
+      nonceTakenErr.name = 'NonceTakenErr'
+      return this.emit('tx:failed', txId, nonceTakenErr)
+    }
+
     // get latest transaction status
     let txParams
     try {
       txParams = await this.query.getTransactionByHash(txHash)
       if (!txParams) return
       if (txParams.blockNumber) {
-        this.emit('txConfirmed', txId)
+        this.emit('tx:confirmed', txId)
       }
     } catch (err) {
       txMeta.warning = {
-        error: err,
+        error: err.message,
         message: 'There was a problem loading this transaction.',
       }
-      this.emit('txWarning', txMeta)
-      throw err
+      this.emit('tx:warning', txMeta, err)
     }
   }
 
@@ -166,4 +171,13 @@ module.exports = class PendingTransactionTracker extends EventEmitter {
     }
     nonceGlobalLock.releaseLock()
   }
+
+  async _checkIfNonceIsTaken (txMeta) {
+    const completed = this.getCompletedTransactions()
+    const sameNonce = completed.filter((otherMeta) => {
+      return otherMeta.txParams.nonce === txMeta.txParams.nonce
+    })
+    return sameNonce.length > 0
+  }
+
 }