+++ mascara/src/proxy.js
@@ -1,9 +1,8 @@
 const createParentStream = require('iframe-stream').ParentStream
 const SWcontroller = require('client-sw-ready-event/lib/sw-client.js')
 const SwStream = require('sw-stream/lib/sw-stream.js')
-const SetupUntrustedComunication = ('./lib/setup-untrusted-connection.js')
 
-let intervalDelay =  Math.floor(Math.random() * (30000 - 1000)) + 1000
+const intervalDelay = Math.floor(Math.random() * (30000 - 1000)) + 1000
 const background = new SWcontroller({
   fileName: '/background.js',
   letBeIdle: false,
@@ -12,8 +11,8 @@ const background = new SWcontroller({
 })
 
 const pageStream = createParentStream()
-background.on('ready', (_) => {
-  let swStream = SwStream({
+background.on('ready', () => {
+  const swStream = SwStream({
     serviceWorker: background.controller,
     context: 'dapp',
   })